backend/requirements.txt
Flask
requests
beautifulsoup4
pytesseract
pillow
lxml
flask-cors
dnspython
